RateDriving/rd/src/requirements.txt
falcon
falcon_cors
psycopg2-binary
gunicorn
pylint
requests